gui/pyproject.toml
[project]
name = "gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cobs>=1.2.1",
    "msgpack>=1.1.0",
    "protobuf>=5.29.3",
    "protobug>=0.2.0",
    "pyqt6>=6.8.0",
    "pyserial>=3.5",
    "pystream>=0.1",
    "pystream-protobuf>=1.6.4",
]
